# hrm-backend/src/main/java/com/pirisa/hrm/controller/AttendanceController.java
package com.pirisa.hrm.controller;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.pirisa.hrm.dto.AttendanceEmployeeDTO;
import com.pirisa.hrm.dto.BulkAttendanceDataDTO;
import com.pirisa.hrm.model.Attendance;
import com.pirisa.hrm.service.AttendanceService;
import com.pirisa.hrm.service.CompanyAccessService;
import com.pirisa.hrm.service.EmployeeService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.core.io.ByteArrayResource;
import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import java.io.IOException;
import java.time.LocalDate;
import java.time.format.DateTimeFormatter;
import java.time.format.DateTimeParseException;
import java.util.Collections;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

@RestController
@RequestMapping("/api/attendance")
public class AttendanceController {

    @Autowired
    private AttendanceService attendanceService;

    @Autowired
    private CompanyAccessService companyAccessService;

    @Autowired
    private EmployeeService employeeService;

    private static final DateTimeFormatter DATE_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd");

    @PostMapping(value = "/bulk-mark", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<?> bulkMarkAttendance(@RequestBody List<Attendance> attendanceList) {
        try {
            List<Attendance> savedEntries = attendanceService.markBulkAttendance(attendanceList);
            Map<String, Object> responseBody = new HashMap<>();
            responseBody.put("resultCode", 100);
            responseBody.put("resultDesc", "Bulk attendance saved successfully");
            responseBody.put("attendanceList", savedEntries);
            return ResponseEntity.ok(responseBody);
        } catch (IllegalArgumentException ex) {
            return ResponseEntity.badRequest()
                    .body(Collections.singletonMap("error", ex.getMessage()));
        } catch (Exception ex) {
            ex.printStackTrace(); // Log the exception
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "Failed to save bulk attendance"));
        }
    }

    @GetMapping(value = "/bulk-data", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<?> getBulkAttendanceData(
            @RequestParam(value = "attendanceDate") String attendanceDateText,
            @RequestParam(value = "companyId") Long companyId,
            @RequestParam(value = "departmentId", required = false) Long departmentId,
            Authentication authentication) {
        if (companyId == null || !companyAccessService.canAccessCompany(authentication.getName(), companyId)) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN)
                    .body(Collections.singletonMap("error", "You cannot access attendance data for this company."));
        }
        try {
            LocalDate attendanceDate = LocalDate.parse(attendanceDateText, DATE_FORMATTER);
            BulkAttendanceDataDTO attendanceData = attendanceService.getBulkAttendanceData(attendanceDate, companyId, departmentId);
            Map<String, Object> responseBody = new HashMap<>();
            responseBody.put("resultCode", 100);
            responseBody.put("resultDesc", "Bulk attendance data fetched successfully");
            responseBody.put("attendanceData", attendanceData);
            return ResponseEntity.ok(responseBody);
        } catch (DateTimeParseException ex) {
            return ResponseEntity.badRequest()
                    .body(Collections.singletonMap("error", "Invalid attendanceDate format. Use yyyy-MM-dd."));
        } catch (Exception ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "Failed to fetch bulk attendance data"));
        }
    }

    @GetMapping(value = "/overview", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<?> getAttendanceOverview(
            @RequestParam(value = "attendanceDate") String attendanceDateText,
            @RequestParam(value = "companyId") Long companyId,
            @RequestParam(value = "departmentId", required = false) Long departmentId,
            Authentication authentication) {
        if (companyId == null || !companyAccessService.canAccessCompany(authentication.getName(), companyId)) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN)
                    .body(Collections.singletonMap("error", "You cannot access attendance data for this company."));
        }
        try {
            LocalDate attendanceDate = LocalDate.parse(attendanceDateText, DATE_FORMATTER);
            BulkAttendanceDataDTO attendanceData = attendanceService.getBulkAttendanceData(attendanceDate, companyId, departmentId);

            Map<String, Object> summary = new HashMap<>();
            summary.put("attendanceDate", attendanceDateText);
            summary.put("companyId", companyId);
            summary.put("departmentId", departmentId);
            summary.put("pendingCount", attendanceData.getPendingEmployees() == null ? 0 : attendanceData.getPendingEmployees().size());
            summary.put("attendedCount", attendanceData.getAttendedEmployees() == null ? 0 : attendanceData.getAttendedEmployees().size());
            summary.put("excludedCount", attendanceData.getExcludedEmployees() == null ? 0 : attendanceData.getExcludedEmployees().size());

            Map<String, Object> responseBody = new HashMap<>();
            responseBody.put("resultCode", 100);
            responseBody.put("resultDesc", "Attendance overview fetched successfully");
            responseBody.put("summary", summary);
            responseBody.put("attendanceData", attendanceData);
            return ResponseEntity.ok(responseBody);
        } catch (DateTimeParseException ex) {
            return ResponseEntity.badRequest()
                    .body(Collections.singletonMap("error", "Invalid attendanceDate format. Use yyyy-MM-dd."));
        } catch (Exception ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "Failed to fetch attendance overview"));
        }
    }

    @GetMapping(value = "/company/{companyId}", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<?> getCompanyAttendanceList(
            @PathVariable long companyId,
            Authentication authentication) {
        if (!companyAccessService.canAccessCompany(authentication.getName(), companyId)) {
            return forbiddenCompanyAttendanceResponse();
        }

        try {
            List<AttendanceEmployeeDTO> employees = employeeService.getAttendanceByCompanyId(companyId);
            if (employees.isEmpty()) {
                return ResponseEntity.status(HttpStatus.NOT_FOUND)
                        .body(Collections.singletonMap("message", "No Attendance List found for this company ID"));
            }

            Map<String, Object> response = new HashMap<>();
            response.put("resultCode", 100);
            response.put("resultDesc", "Successful");
            response.put("EmployeeList", employees);
            return ResponseEntity.ok(response);
        } catch (Exception ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "An error occurred while fetching company attendance"));
        }
    }

    @GetMapping(value = "/company/{companyId}/latest", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<?> getLatestCompanyAttendanceList(
            @PathVariable long companyId,
            Authentication authentication) {
        if (!companyAccessService.canAccessCompany(authentication.getName(), companyId)) {
            return forbiddenCompanyAttendanceResponse();
        }

        try {
            List<AttendanceEmployeeDTO> employees = employeeService.getLastAttendanceByCompanyId(companyId);
            Map<String, Object> response = new HashMap<>();
            response.put("resultCode", 100);
            response.put("resultDesc", "Successful");
            response.put("EmployeeList", employees);
            return ResponseEntity.ok(response);
        } catch (Exception ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "An error occurred while fetching latest company attendance"));
        }
    }

    @GetMapping(value = "/company/{companyId}/month/{month}", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<?> getCompanyAttendanceByMonth(
            @PathVariable long companyId,
            @PathVariable int month,
            Authentication authentication) {
        if (month < 1 || month > 12) {
            return ResponseEntity.badRequest()
                    .body(Collections.singletonMap("error", "Month must be between 1 and 12."));
        }
        if (!companyAccessService.canAccessCompany(authentication.getName(), companyId)) {
            return forbiddenCompanyAttendanceResponse();
        }

        try {
            List<AttendanceEmployeeDTO> employees = employeeService.getAttendanceByCompanyIdAndMonth(companyId, month);
            Map<String, Object> response = new HashMap<>();
            response.put("resultCode", 100);
            response.put("resultDesc", "Successful");
            response.put("EmployeeList", employees);
            return ResponseEntity.ok(response);
        } catch (Exception ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "An error occurred while fetching monthly company attendance"));
        }
    }

    private ResponseEntity<?> forbiddenCompanyAttendanceResponse() {
        return ResponseEntity.status(HttpStatus.FORBIDDEN)
                .body(Collections.singletonMap("error", "You do not have access to this company attendance data."));
    }

    @PostMapping(value = "/import-excel", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
    public ResponseEntity<?> importAttendanceFromExcel(
            @RequestPart("file") MultipartFile file,
            @RequestParam(value = "createdBy", required = false) String createdBy) {
        try {
            List<Attendance> importedRecords = attendanceService.importAttendanceFromExcel(file, createdBy);
            Map<String, Object> responseBody = new HashMap<>();
            responseBody.put("resultCode", 100);
            responseBody.put("resultDesc", "Attendance imported successfully");
            responseBody.put("importedCount", importedRecords.size());
            responseBody.put("attendanceList", importedRecords);
            return ResponseEntity.ok(responseBody);
        } catch (IllegalArgumentException ex) {
            return ResponseEntity.badRequest()
                    .body(Collections.singletonMap("error", ex.getMessage()));
        } catch (IOException ex) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                    .body(Collections.singletonMap("error", "Invalid Excel file format"));
        } catch (Exception ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "Failed to import attendance from Excel"));
        }
    }

    @GetMapping(value = "/download-excel")
    public ResponseEntity<?> downloadAttendanceExcel(
            @RequestParam(value = "type", required = false) String type,
            @RequestParam(value = "department", required = false) Long departmentId,
            @RequestParam(value = "empId", required = false) Long empId,
            @RequestParam(value = "startDate", required = false) String startDateText,
            @RequestParam(value = "endDate", required = false) String endDateText,
            Authentication authentication) {
        long companyId = companyAccessService.getCompanyIdForUser(authentication.getName());
        if (companyId <= 0) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN)
                    .body(Collections.singletonMap("error", "Unable to determine your company for this report."));
        }
        try {
            LocalDate startDate = parseDateOrNull(startDateText);
            LocalDate endDate = parseDateOrNull(endDateText);
            if (type != null && startDate != null) {
                LocalDate[] dateRange = buildDateRange(type, startDate);
                startDate = dateRange[0];
                endDate = dateRange[1];
            }
            byte[] excelBytes = attendanceService.exportAttendanceToExcel(
                    companyId, departmentId, empId, startDate, endDate);
            String filename = buildExcelFilename(type, departmentId, empId, startDate, endDate);
            ByteArrayResource resource = new ByteArrayResource(excelBytes);

            return ResponseEntity.ok()
                    .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + filename + "\"")
                    .contentType(MediaType.parseMediaType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"))
                    .contentLength(excelBytes.length)
                    .body(resource);
        } catch (DateTimeParseException ex) {
            return ResponseEntity.badRequest()
                    .body(Collections.singletonMap("error", "Invalid date format. Use yyyy-MM-dd."));
        } catch (IllegalArgumentException ex) {
            return ResponseEntity.badRequest()
                    .body(Collections.singletonMap("error", ex.getMessage()));
        } catch (IOException ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "Unable to generate Excel report"));
        } catch (Exception ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(Collections.singletonMap("error", "Unexpected error while downloading attendance report"));
        }
    }

    private LocalDate parseDateOrNull(String dateText) {
        if (dateText == null || dateText.isBlank()) {
            return null;
        }
        return LocalDate.parse(dateText, DATE_FORMATTER);
    }

    private LocalDate[] buildDateRange(String type, LocalDate anchorDate) {
        if (type == null) {
            return new LocalDate[]{anchorDate, anchorDate};
        }
        switch (type.trim().toUpperCase()) {
            case "DAILY":
                return new LocalDate[]{anchorDate, anchorDate};
            case "WEEKLY":
                return new LocalDate[]{anchorDate, anchorDate.plusDays(6)};
            case "MONTHLY":
                return new LocalDate[]{anchorDate.withDayOfMonth(1), anchorDate.withDayOfMonth(anchorDate.lengthOfMonth())};
            case "YEARLY":
                return new LocalDate[]{anchorDate.withDayOfYear(1), anchorDate.withDayOfYear(anchorDate.lengthOfYear())};
            default:
                return new LocalDate[]{anchorDate, anchorDate};
        }
    }

    private String buildExcelFilename(String type, Long departmentId, Long empId, LocalDate startDate, LocalDate endDate) {
        StringBuilder name = new StringBuilder("attendance-report");
        if (type != null && !type.isBlank()) {
            name.append("-").append(type.toLowerCase());
        }
        if (departmentId != null) {
            name.append("-dept-").append(departmentId);
        }
        if (empId != null) {
            name.append("-emp-").append(empId);
        }
        if (startDate != null) {
            name.append("-").append(startDate.format(DATE_FORMATTER));
        }
        if (endDate != null && !endDate.equals(startDate)) {
            name.append("-to-").append(endDate.format(DATE_FORMATTER));
        }
        name.append(".xlsx");
        return name.toString();
    }

    @ExceptionHandler(Exception.class)
    public ResponseEntity<String> handleException(Exception e) {
        e.printStackTrace();
        Map<String, Object> errorResponse = new HashMap<>();
        errorResponse.put("resultCode", 101);
        errorResponse.put("resultDesc", "ERROR");

        String jsonResponse;
        try {
            jsonResponse = new ObjectMapper().writeValueAsString(errorResponse);
        } catch (Exception ex) {
            jsonResponse = "{\"resultCode\":101,\"resultDesc\":\"ERROR\"}";
        }
        return new ResponseEntity<>(jsonResponse, HttpStatus.INTERNAL_SERVER_ERROR);
    }

    @PutMapping(value = "/clock-out/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<?> clockOutAttendance(@PathVariable("id") Long id, @RequestBody Map<String, Object> payload) {
        try {
            String endedAt = payload.get("endedAt") != null ? String.valueOf(payload.get("endedAt")) : null;
            String departureReason = payload.get("departureReason") != null ? String.valueOf(payload.get("departureReason")) : null;
            String departureNotes = payload.get("departureNotes") != null ? String.valueOf(payload.get("departureNotes")) : null;

            Attendance updated = attendanceService.clockOutAttendance(id, endedAt, departureReason, departureNotes);
            Map<String, Object> responseBody = new HashMap<>();
            responseBody.put("resultCode", 100);
            responseBody.put("resultDesc", "Clock-out updated successfully");
            responseBody.put("attendance", updated);
            return ResponseEntity.ok(responseBody);
        } catch (IllegalArgumentException ex) {
            return ResponseEntity.badRequest().body(Collections.singletonMap("error", ex.getMessage()));
        } catch (Exception ex) {
            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(Collections.singletonMap("error", "Failed to clock out attendance"));
        }
    }
}